+++ example/config.js
@@ -2,7 +2,7 @@ const { createConfig } = require('larana-js')
 
 module.exports = createConfig({
 	defaultTheme: 'dark',
-	debug: true,
+	debug: false,
 	debugOptions: {
 		renderOutline: true,
 		renderPaddings: true,