+++ src/ui/components/text.js
@@ -1,3 +1,4 @@
+const { t } = require('../shapes')
 const { BaseComponent } = require('./base')
 const { figure } = require('./figure')
 
@@ -13,12 +14,6 @@ class TextComponent extends BaseComponent {
 		this.text = text
 	}
 
-	clearText(text) {
-		return String(text)
-			.replaceAll('"', '\\"')
-			.replaceAll("'", "\\'")
-	}
-
 	root() {
 		const { modelValue } = this.useModel()
 		const style = this.computeStyle()
@@ -26,17 +21,17 @@ class TextComponent extends BaseComponent {
 		return figure({
 			template: (fig, queue) => {
 				const d = fig.computeDimensions()
-				queue.add('text', {
-					// x: d.x,
-					// y: d.y,
+
+				const r = t({
 					x: d.x + (d.w / 2),
 					y: d.y + (d.h / 2),
+					text: modelValue,
+					...style,
 					font: style.font,
-					text: this.clearText(modelValue),
-					align: style.textAlign,
-					baseline: style.textBaseline,
-					fg: style.fg ?? '#000',
+					fg: style.fg,
 				})
+
+				r.to(queue)
 			},
 		})
 	}