+++ src/ui/shapes/index.js
@@ -2,6 +2,8 @@ module.exports = {
 	...require('./shape.js'),
 	...require('./rect.js'),
 	...require('./line.js'),
+	...require('./arrow.js'),
 	...require('./point.js'),
 	...require('./arc.js'),
+	...require('./t.js'),
 }